requirements.txt
pandas
openpyxl
requests
plotly
numpy
python-dotenv
ollama